Whew.
Needs more UAT, review project rules/stipulations, readme update, nomenclature sanitization, decide what to do with test project, UI cleanup and endpoint cleanup. IDE recommendations started getting a little out of control, need to revisit some of the "shortening" it did with the methods/method param names; not really a fan.

Not bad for on and off the PC for a day.